Move 'MarkedDates' type to types; Move `VelocityTracker` to Agenda directory; Move all tools to util directory; Moving 'interface' tools to 'dateutils';
