+++ src/calendar-list/new.tsx
@@ -2,8 +2,8 @@ import XDate from 'xdate';
 import React, {useCallback, useEffect, useRef, useState, useMemo} from 'react';
 import {View, ScrollViewProps, ScrollView} from 'react-native';
 import constants from '../commons/constants';
-import {toMarkingFormat} from '../interface';
-import {extractHeaderProps} from '../componentUpdater';
+import {toMarkingFormat} from '../utils/dateutils';
+import {extractHeaderProps} from '../utils/componentUpdater';
 import Calendar, {CalendarProps} from '../calendar';
 import CalendarHeader from '../calendar/header';
 import InfiniteList from '../infinite-list';
