+++ src/expandableCalendar/WeekCalendar/new.tsx
@@ -8,11 +8,10 @@ import WeekDaysNames from '../../commons/WeekDaysNames';
 import {CalendarListProps} from '../../calendar-list';
 import CalendarContext from '../../expandableCalendar/Context';
 import styleConstructor from '../style';
-import {toMarkingFormat} from '../../interface';
-import {extractCalendarProps} from '../../componentUpdater';
+import {extractCalendarProps} from '../../utils/componentUpdater';
 import constants from '../../commons/constants';
 import {UpdateSources} from '../commons';
-import {sameWeek} from '../../dateutils';
+import {sameWeek, toMarkingFormat} from '../../utils/dateutils';
 import {DateData} from '../../types';
 
 export interface WeekCalendarProps extends CalendarListProps {
