+++ src/expandableCalendar/agendaList.tsx
@@ -22,9 +22,8 @@ import {
   ViewToken
 } from 'react-native';
 
-import {isToday, isGTE, sameDate} from '../dateutils';
-import {getMoment} from '../momentResolver';
-import {parseDate} from '../interface';
+import {isToday, isGTE, sameDate, parseDate} from '../utils/dateutils';
+import {getMoment} from '../utils/momentResolver';
 import {getDefaultLocale} from '../services';
 import {UpdateSources, todayString} from './commons';
 import {Theme} from '../types';
