+++ src/index.ts
@@ -22,4 +22,4 @@ export type {TimelineListProps, TimelineListRenderItemInfo} from './timeline-lis
 export {default as CalendarUtils} from './services';
 export type {DateData, AgendaEntry, AgendaSchedule} from './types';
 export {default as LocaleConfig} from 'xdate';
-export {default as Profiler} from './Profiler';
+export {default as Profiler} from './utils/Profiler';
