+++ src/utils/__tests__/componentUpdater.spec.js
@@ -1,4 +1,4 @@
-import {shouldUpdate} from './componentUpdater';
+import {shouldUpdate} from '../componentUpdater';
 
 describe('component updater', () => {
   it('should return true if two different objects are provided with same path', async () => {
