+++ src/utils/dateutils.ts
@@ -1,6 +1,4 @@
 const XDate = require('xdate');
-const {parseDate, toMarkingFormat} = require('./interface');
-const {getDefaultLocale} = require('./services');
 
 const latinNumbersPattern = /[0-9]/g;
 
@@ -62,7 +60,7 @@ export function isLTE(a: XDate, b: XDate) {
 }
 
 export function formatNumbers(date: any) {
-  const numbers = getDefaultLocale().numbers;
+  const numbers = getLocale().numbers;
   return numbers ? date.toString().replace(latinNumbersPattern, (char: any) => numbers[+char]) : date;
 }
 
@@ -89,7 +87,7 @@ export function month(date: XDate) {
 }
 
 export function weekDayNames(firstDayOfWeek = 0) {
-  let weekDaysNames = getDefaultLocale().dayNamesShort;
+  let weekDaysNames = getLocale().dayNamesShort;
   const dayShift = firstDayOfWeek % 7;
   if (dayShift) {
     weekDaysNames = weekDaysNames.slice(dayShift).concat(weekDaysNames.slice(0, dayShift));
@@ -188,4 +186,60 @@ export function getPartialWeekDates(date: string | undefined, numberOfDays = 7) 
 export function generateDay(originDate: string, daysOffset = 0) {
   const baseDate = new XDate(originDate);
   return toMarkingFormat(baseDate.clone().addDays(daysOffset));
+}
+
+export function padNumber(n: number) {
+  if (n < 10) {
+    return '0' + n;
+  }
+  return n;
+}
+
+export function xdateToData(d: XDate) {
+  const dateString = toMarkingFormat(d);
+  return {
+    year: d.getFullYear(),
+    month: d.getMonth() + 1,
+    day: d.getDate(),
+    timestamp: new XDate(dateString, true).getTime(),
+    dateString: dateString
+  };
+}
+
+export function parseDate(d?: any) {
+  if (!d) {
+    return;
+  } else if (d.timestamp) {
+    // conventional data timestamp
+    return new XDate(d.timestamp, true);
+  } else if (d instanceof XDate) {
+    // xdate
+    return new XDate(toMarkingFormat(d), true);
+  } else if (d.getTime) {
+    // javascript date
+    const dateString = d.getFullYear() + '-' + padNumber(d.getMonth() + 1) + '-' + padNumber(d.getDate());
+    return new XDate(dateString, true);
+  } else if (d.year) {
+    const dateString = d.year + '-' + padNumber(d.month) + '-' + padNumber(d.day);
+    return new XDate(dateString, true);
+  } else if (d) {
+    // timestamp number or date formatted as string
+    return new XDate(d, true);
+  }
+}
+
+export function toMarkingFormat(d: XDate) {
+  if (!isNaN(d.getTime())) {
+    const year = `${d.getFullYear()}`;
+    const month = d.getMonth() + 1;
+    const doubleDigitMonth = month < 10 ? `0${month}` : `${month}`;
+    const day = d.getDate();
+    const doubleDigitDay = day < 10 ? `0${day}` : `${day}`;
+    return year + '-' + doubleDigitMonth + '-' + doubleDigitDay;
+  }
+  return 'Invalid Date';
+}
+
+export function getLocale() {
+  return XDate.locales[XDate.defaultLocale];
 }
